+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/SerializedRenderPipelineSettings.cs
@@ -34,7 +34,6 @@ namespace UnityEditor.Rendering.HighDefinition
         public SerializedProperty supportCustomPass;
         public SerializedProperty customBufferFormat;
         public SerializedScalableSetting planarReflectionResolution;
-
         public SerializedProperty supportDecals;
         public SerializedProperty supportDecalLayers;
         public SerializedProperty supportSurfaceGradient;
@@ -55,7 +54,6 @@ namespace UnityEditor.Rendering.HighDefinition
         internal SerializedProperty probeVolumeTextureSize;
         internal SerializedProperty supportProbeVolumeStreaming;
         internal SerializedProperty probeVolumeSHBands;
-
 
         public SerializedGlobalLightLoopSettings lightLoopSettings;
         public SerializedHDShadowInitParameters hdShadowInitParams;
@@ -113,7 +111,6 @@ namespace UnityEditor.Rendering.HighDefinition
             probeVolumeTextureSize = root.Find((RenderPipelineSettings s) => s.probeVolumeMemoryBudget);
             supportProbeVolumeStreaming = root.Find((RenderPipelineSettings s) => s.supportProbeVolumeStreaming);
             probeVolumeSHBands = root.Find((RenderPipelineSettings s) => s.probeVolumeSHBands);
-
             supportRayTracing = root.Find((RenderPipelineSettings s) => s.supportRayTracing);
             supportedRayTracingMode = root.Find((RenderPipelineSettings s) => s.supportedRayTracingMode);